World loader just reads from png now
I decided to make it so that World loader will get the data from the png directly this saves space and completly aviods having to create a whole new file. I din't realize i could do this before and now i do. All maps will now be in the Maps folder in res.