added anitmaion
animation is now added when we make a sprite sheet with animations we make a .data file along with it that tells id,u,v,width and height of the sprite. to create animation you first load up the spritesheetdata from the spritesheetloader class then you pass that along with the model,fps,start and end frames. then use .update() every frame to update the animation.