FIxing guis
You can now add,remove and replace Elements.I also created a stack for storing previos states so that when you press the back button it will just pop the last state off the stack. Allso to exit the game(only in overworld) you press the escape key and then select quit. Still need to fix some of the movement for the arrow though.Almost ready to start more battle stuff now that the guis are done i just need to implement the functions.