Working on Enemy ai(healing)
so far when enemy's hp is below 25% the enemy will use it's best healing item or move. Later I will add a chance to use a attack move if it has one that can kill the player instead of healing. All of this is just geniric AI later we will add more enemies types that extend this class and overide the takeTurn() method.